+++ lib/tynn/secure_headers.rb
@@ -31,15 +31,13 @@ class Tynn
   # [xss]: https://www.owasp.org/index.php/Cross-site_Scripting_(XSS)
   #
   module SecureHeaders
-    HEADERS = {
-      "X-Content-Type-Options" => "nosniff",
-      "X-Frame-Options" => "SAMEORIGIN",
-      "X-Permitted-Cross-Domain-Policies" => "none",
-      "X-XSS-Protection" => "1; mode=block"
-    } # :nodoc:
-
-    def default_headers # :nodoc:
-      return super.merge(HEADERS)
+    def self.setup(app) # :nodoc:
+      app.settings[:default_headers].update(
+        "X-Content-Type-Options" => "nosniff",
+        "X-Frame-Options" => "SAMEORIGIN",
+        "X-Permitted-Cross-Domain-Policies" => "none",
+        "X-XSS-Protection" => "1; mode=block"
+      )
     end
   end
 end