+++ test/secure_headers_test.rb
@@ -12,9 +12,16 @@ test "secure headers" do
   app = Tynn::Test.new
   app.get("/")
 
+  secure_headers = {
+    "X-Content-Type-Options" => "nosniff",
+    "X-Frame-Options" => "SAMEORIGIN",
+    "X-Permitted-Cross-Domain-Policies" => "none",
+    "X-XSS-Protection" => "1; mode=block"
+  }
+
   headers = app.res.headers
 
-  Tynn::SecureHeaders::HEADERS.each do |header, value|
+  secure_headers.each do |header, value|
     assert_equal(value, headers[header])
   end
 end